ch(#160829827): validate request params
