+++ src/helpers.js
@@ -1,0 +1,9 @@
+export const jsonParse = arg => {
+  let parsed;
+  try {
+    parsed = JSON.parse(arg);
+  } catch (e) {
+    console.log(e);
+  }
+  return parsed;
+};
